+++ src/test/java/ibm/jceplus/junit/openjceplus/TestAll.java
@@ -24,12 +24,6 @@ import org.junit.platform.suite.api.Suite;
     TestAliases.class,
     TestAttributes.class,
     TestByteArrayOutputDelay.class,
-    TestChaCha20.class,
-    TestChaCha20KAT.class,
-    TestChaCha20NoReuse.class,
-    TestChaCha20Poly1305.class,
-    TestChaCha20Poly1305ByteBuffer.class,
-    TestChaCha20Poly1305ChunkUpdate.class,
     TestDeterministic.class,
     TestDH.class,
     TestDHInteropSunJCE.class,
