+++ src/test/java/ibm/jceplus/junit/tests/TestChaCha20.java
@@ -1,15 +1,14 @@
 /*
- * Copyright IBM Corp. 2023, 2025
+ * Copyright IBM Corp. 2023, 2026
  *
  * This code is free software; you can redistribute it and/or modify it
  * under the terms provided by IBM in the LICENSE file that accompanied
  * this code, including the "Classpath" Exception described therein.
  */
 
-package ibm.jceplus.junit.base;
+package ibm.jceplus.junit.tests;
 
 import com.ibm.crypto.plus.provider.ChaCha20Constants;
-import java.security.AlgorithmParameters;
 import java.security.InvalidAlgorithmParameterException;
 import java.security.InvalidKeyException;
 import java.security.NoSuchAlgorithmException;
@@ -18,16 +17,29 @@ import java.util.Arrays;
 import javax.crypto.Cipher;
 import javax.crypto.KeyGenerator;
 import javax.crypto.SecretKey;
+import javax.crypto.SecretKeyFactory;
 import javax.crypto.ShortBufferException;
+import javax.crypto.spec.ChaCha20ParameterSpec;
 import javax.crypto.spec.IvParameterSpec;
 import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.TestInstance;
+import org.junit.jupiter.params.Parameter;
+import org.junit.jupiter.params.ParameterizedClass;
+import org.junit.jupiter.params.provider.MethodSource;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.junit.jupiter.api.Assertions.fail;
 
-public class BaseTestChaCha20Poly1305 extends BaseTestCipher implements ChaCha20Constants {
+@Tag(Tags.OPENJCEPLUS_NAME)
+@TestInstance(TestInstance.Lifecycle.PER_CLASS)
+@ParameterizedClass
+@MethodSource("ibm.jceplus.junit.tests.TestArguments#getOpenJCEPlusOnly")
+public class TestChaCha20 extends BaseTest implements ChaCha20Constants {
 
+    @Parameter(0)
+    TestProvider provider;
 
     // 14 bytes: PASSED
     static final byte[] PLAIN_TEXT_14 = "12345678123456".getBytes();
@@ -53,25 +65,34 @@ public class BaseTestChaCha20Poly1305 extends BaseTestCipher implements ChaCha20
     static final byte[] NONCEA_12_BYTE = "012345678123".getBytes();
     static final byte[] NONCE_13_BYTE = "1234567812345".getBytes();
 
-    static final byte[] BAD_TAG_16 = "BaadTaagBaadTaag".getBytes();
+    static final int COUNTER_0 = 0;
+    static final int COUNTER_1 = 1;
+    static final int COUNTER_MIN = Integer.MIN_VALUE;
+    static final int COUNTER_MAX = Integer.MAX_VALUE;
 
-    static final byte[] CHACHA20_POLY1305_AAD = "12345".getBytes(); //"ChaCha20-Poly1305 AAD".getBytes();
-
-    static final String CHACHA20_POLY1305_ALGORITHM = "ChaCha20-Poly1305";
     static final String CHACHA20_ALGORITHM = "ChaCha20";
 
-    static final IvParameterSpec CHACHA20_POLY1305_PARAM_SPEC = new IvParameterSpec(NONCE_12_BYTE);
+    static final ChaCha20ParameterSpec CHACHA20_PARAM_SPEC_COUNTER_0 = new ChaCha20ParameterSpec(
+            NONCE_12_BYTE, COUNTER_0);
+    static final ChaCha20ParameterSpec CHACHA20_PARAM_SPEC_COUNTER_1 = new ChaCha20ParameterSpec(
+            NONCE_12_BYTE, COUNTER_1);
+    static final ChaCha20ParameterSpec CHACHA20_PARAM_SPEC_COUNTER_MIN = new ChaCha20ParameterSpec(
+            NONCE_12_BYTE, COUNTER_MIN);
+    static final ChaCha20ParameterSpec CHACHA20_PARAM_SPEC_COUNTER_MAX = new ChaCha20ParameterSpec(
+            NONCE_12_BYTE, COUNTER_MAX);
 
 
     protected KeyGenerator keyGen = null;
     protected SecretKey key = null;
-    protected IvParameterSpec paramSpec = null;
+    protected SecretKeyFactory keyFactory = null;
+    protected ChaCha20ParameterSpec paramSpec = null;
     protected Cipher cp = null;
     protected boolean success = true;
     protected int specifiedKeySize = 0;
 
     @BeforeEach
     public void setUp() throws Exception {
+        setAndInsertProvider(provider);
         keyGen = KeyGenerator.getInstance(CHACHA20_ALGORITHM, getProviderName());
         if (specifiedKeySize > 0) {
             keyGen.init(specifiedKeySize);
@@ -80,24 +101,33 @@ public class BaseTestChaCha20Poly1305 extends BaseTestCipher implements ChaCha20
     }
 
     @Test
-    public void testChaCha20Poly1305IllegalKeyNonceReuse() throws Exception {
+    public void testChaCha20KeyFactory() throws Exception {
+
         try {
-            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
-            cp.init(Cipher.ENCRYPT_MODE, key, CHACHA20_POLY1305_PARAM_SPEC);
-            cp.init(Cipher.DECRYPT_MODE, key, CHACHA20_POLY1305_PARAM_SPEC);
-
-            //fail("Expected InvalidKeyException did not occur");
-
-        } catch (InvalidKeyException e) {
-            e.printStackTrace();
-            throw e;
+            keyFactory = SecretKeyFactory.getInstance(CHACHA20_ALGORITHM, getProviderName());
+        } catch (Exception e) {
+            fail("Got unexpected exception on ChaCha20KeyFactory.getInstance()...");
         }
     }
 
     @Test
-    public void testChaCha20Poly1305ShortBuffer() throws Exception {
+    public void testChaCha20IllegalKeyNonceReuse() throws Exception {
         try {
-            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
+            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
+            cp.init(Cipher.ENCRYPT_MODE, key, CHACHA20_PARAM_SPEC_COUNTER_0);
+            cp.init(Cipher.DECRYPT_MODE, key, CHACHA20_PARAM_SPEC_COUNTER_0);
+
+            //fail("Expected InvalidKeyException did not occur");
+
+        } catch (InvalidKeyException e) {
+            assertTrue(false);
+        }
+    }
+
+    @Test
+    public void testChaCha20ShortBuffer() throws Exception {
+        try {
+            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
 
             // Encrypt the plain text
             cp.init(Cipher.ENCRYPT_MODE, key);
@@ -111,13 +141,13 @@ public class BaseTestChaCha20Poly1305 extends BaseTestCipher implements ChaCha20
 
     @Test
     public void testChaCha20Poly1305EncryptParamsAfterShortBufferRetry() throws Exception {
-        cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
-        IvParameterSpec ivParamSpec = new IvParameterSpec(NONCE_12_BYTE);
+        cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
+        ChaCha20ParameterSpec paraSpec = new ChaCha20ParameterSpec(NONCE_12_BYTE, COUNTER_0);
 
         // Try encrypting with a short buffer.
         try {
-            cp.init(Cipher.ENCRYPT_MODE, key, ivParamSpec);
-            byte[] cipherText = new byte[PLAIN_TEXT.length + 15];
+            cp.init(Cipher.ENCRYPT_MODE, key, paraSpec);
+            byte[] cipherText = new byte[PLAIN_TEXT.length - 1];
             cp.doFinal(PLAIN_TEXT, 0, PLAIN_TEXT.length, cipherText);
             fail("Expected ShortBufferException did not occur");
         } catch (ShortBufferException sbe) {
@@ -134,20 +164,20 @@ public class BaseTestChaCha20Poly1305 extends BaseTestCipher implements ChaCha20
         }
 
         // Retry with a larger buffer after re-initializing.
-        ivParamSpec = new IvParameterSpec(NONCEA_12_BYTE);
-        cp.init(Cipher.ENCRYPT_MODE, key, ivParamSpec);
+        paraSpec = new ChaCha20ParameterSpec(NONCEA_12_BYTE, COUNTER_0);
+        cp.init(Cipher.ENCRYPT_MODE, key, paraSpec);
         byte[] largerCipherTextBuffer = new byte[PLAIN_TEXT.length + 16];
         cp.doFinal(PLAIN_TEXT, 0, PLAIN_TEXT.length, largerCipherTextBuffer);
     }
 
     @Test
     public void testChaCha20Poly1305EncryptNoParamsAfterShortBufferRetry() throws Exception {
-        cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
+        cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
 
         // Try encrypting with a short buffer.
         try {
             cp.init(Cipher.ENCRYPT_MODE, key);
-            byte[] cipherText = new byte[PLAIN_TEXT.length + 15];
+            byte[] cipherText = new byte[PLAIN_TEXT.length - 1];
             cp.doFinal(PLAIN_TEXT, 0, PLAIN_TEXT.length, cipherText);
             fail("Expected ShortBufferException did not occur");
         } catch (ShortBufferException sbe) {
@@ -170,62 +200,26 @@ public class BaseTestChaCha20Poly1305 extends BaseTestCipher implements ChaCha20
     }
 
     @Test
-    public void testChaCha20Poly1305DecryptAfterShortBufferRetry() throws Exception {
-        byte[] cipherText = null;
-        Cipher cpl = null;
-        try {
-            cpl = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
-            IvParameterSpec ivParamSpec = new IvParameterSpec(NONCEA_12_BYTE);
-
-            // Encrypt the plain text
-            cpl.init(Cipher.ENCRYPT_MODE, key, ivParamSpec);
-            cipherText = cpl.doFinal(PLAIN_TEXT, 0, PLAIN_TEXT.length);
-
-            ivParamSpec = cpl.getParameters().getParameterSpec(IvParameterSpec.class);
-
-            cpl = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
-            cpl.init(Cipher.DECRYPT_MODE, key, ivParamSpec);
-            byte[] sbPlainText = new byte[15];
-            System.out.println("cipherText.length=" + cipherText.length);
-            System.out.println("sbPlainText.length=" + sbPlainText.length);
-            cpl.doFinal(cipherText, 0, cipherText.length, sbPlainText, 0);
-            fail("Failed to get ShortedBufferException");
-
-        } catch (ShortBufferException ex) {
-            System.out.println("Try retry with a larger buffer");
-        }
-        // try retry with a larger buffer
-        try {
-            byte[] lbPlainTextBuffer = new byte[PLAIN_TEXT.length];
-            cpl.doFinal(cipherText, 0, cipherText.length, lbPlainTextBuffer, 0);
-            assertTrue(Arrays.equals(PLAIN_TEXT, lbPlainTextBuffer));
-        } catch (Exception ex) {
-            ex.printStackTrace();
-            fail("Retying with larger buffer should have worked  with a larger buffer");
-        }
-    }
-
-    @Test
-    public void testChaCha20Poly1305GetBlockSizeEncryptDecrypt() throws Exception {
+    public void testChaCha20GetBlockSizeEncryptDecrypt() throws Exception {
         chaCha20GetBlockSize(Cipher.ENCRYPT_MODE);
         chaCha20GetBlockSize(Cipher.DECRYPT_MODE);
     }
 
 
     public void chaCha20GetBlockSize(int opMode) throws Exception {
-        IvParameterSpec chaCha20ParamSpec = new IvParameterSpec(NONCE_12_BYTE);
-        cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
+        ChaCha20ParameterSpec chaCha20ParamSpec = new ChaCha20ParameterSpec(NONCE_12_BYTE, 0);
+        cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
         cp.init(opMode, key, chaCha20ParamSpec);
         assertTrue((cp.getBlockSize() == ChaCha20_BLOCK_SIZE), "ChaCha20 Block size must be: " + ChaCha20_BLOCK_SIZE);
     }
 
     @Test
-    public void testChaCha20Poly1305ValidTransformations() throws Exception {
+    public void testChaCha20ValidTransformations() throws Exception {
         String transformation = null;
         try {
-            transformation = CHACHA20_POLY1305_ALGORITHM;
+            transformation = CHACHA20_ALGORITHM;
             cp = Cipher.getInstance(transformation, getProviderName());
-            transformation = "ChaCha20-Poly1305/None/NoPadding";
+            transformation = "ChaCha20/None/NoPadding";
             cp = Cipher.getInstance(transformation, getProviderName());
         } catch (NoSuchAlgorithmException ex) {
             fail("NoSuchAlgorithmException occurred for transform: " + transformation);
@@ -233,8 +227,8 @@ public class BaseTestChaCha20Poly1305 extends BaseTestCipher implements ChaCha20
     }
 
     @Test
-    public void testChaCha20Poly1305InvalidTransformation() throws Exception {
-        String transformation = "BogusChaCha20-Poly1305/BogusMode/BogusPadding";
+    public void testChaCha20InvalidTransformation() throws Exception {
+        String transformation = "BogusChaCha20/BogusMode/BogusPadding";
         try {
             cp = Cipher.getInstance(transformation, getProviderName());
             fail("Expected NoSuchAlgorithmException did not occur");
@@ -244,8 +238,8 @@ public class BaseTestChaCha20Poly1305 extends BaseTestCipher implements ChaCha20
     }
 
     @Test
-    public void testChaCha20Poly1305NullKey() throws Exception {
-        cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
+    public void testChaCha20NullKey() throws Exception {
+        cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
         SecretKey nullKey = null;
 
         try {
@@ -255,6 +249,7 @@ public class BaseTestChaCha20Poly1305 extends BaseTestCipher implements ChaCha20
         }
 
         try {
+            //cp.init(Cipher.ENCRYPT_MODE, nullKey, SecureRandom.getInstance("SecureRandom"));
             cp.init(Cipher.ENCRYPT_MODE, nullKey, SecureRandom.getInstance("SHA2DRBG"));
             fail("Expected InvalidKeyException did not occur");
         } catch (InvalidKeyException e) {
@@ -262,151 +257,89 @@ public class BaseTestChaCha20Poly1305 extends BaseTestCipher implements ChaCha20
     }
 
     @Test
-    public void testChaCha20Poly1305NoParamSpec() throws Exception {
+    public void testChaCha20NoParamSpec() throws Exception {
 
         try {
-            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
+            // For ChaCha20 you need to pass in the nonce/iv because
+            // there is no way to get the iv out from this cipher.
+            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
             cp.init(Cipher.ENCRYPT_MODE, key);
-            byte[] cipherText = cp.doFinal(PLAIN_TEXT);
-
-            paramSpec = cp.getParameters().getParameterSpec(IvParameterSpec.class);
-
-            // Verify the text
-            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
-            cp.init(Cipher.DECRYPT_MODE, key, paramSpec);
-            byte[] newPlainText = cp.doFinal(cipherText, 0, cipherText.length);
-
-            assertTrue(Arrays.equals(PLAIN_TEXT, newPlainText));
-
-        } catch (Exception e) {
-            fail("Got unexpected exception on encrypt/decrypt...");
-        }
-    }
-
-    @Test
-    public void testChaCha20Poly1305NullParamSpec() throws Exception {
-
-        try {
-            IvParameterSpec ivSpec = null;
-            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
-            cp.init(Cipher.ENCRYPT_MODE, key, ivSpec);
-            byte[] cipherText = cp.doFinal(PLAIN_TEXT);
-
-            paramSpec = cp.getParameters().getParameterSpec(IvParameterSpec.class);
-
-            // Verify the text
-            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
-            cp.init(Cipher.DECRYPT_MODE, key, paramSpec);
-            byte[] newPlainText = cp.doFinal(cipherText, 0, cipherText.length);
-
-            assertTrue(Arrays.equals(PLAIN_TEXT, newPlainText));
-
-        } catch (Exception e) {
-            fail("Got unexpected exception on encrypt/decrypt...");
-        }
-    }
-
-    @Test
-    public void testChaCha20Poly1305InvalidParamSpec() throws Exception {
-
-        try {
-            byte[] iv = null;
-            IvParameterSpec ivSpec = new IvParameterSpec(iv);
-            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
-            cp.init(Cipher.ENCRYPT_MODE, key, ivSpec);
-            fail("Expected NullPointerException or InvalidAlgorithmParameterException");
-        } catch (NullPointerException npe) {
-        } catch (InvalidAlgorithmParameterException e) {
-        }
-
-        try {
-            byte[] iv = null;
-            IvParameterSpec ivSpec = new IvParameterSpec(iv);
-            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
-            cp.init(Cipher.DECRYPT_MODE, key, ivSpec);
-            fail("Expected NullPointerException or InvalidAlgorithmParameterException");
-        } catch (NullPointerException npe) {
-        } catch (InvalidAlgorithmParameterException e) {
-        }
-
-        try {
-            AlgorithmParameters algParameters = AlgorithmParameters
-                    .getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
-            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
-            cp.init(Cipher.ENCRYPT_MODE, key, algParameters);
-            fail("Expected NullPointerException or InvalidAlgorithmParameterException");
-        } catch (NullPointerException npe) {
-        } catch (InvalidAlgorithmParameterException e) {
-        }
-
-        try {
-            AlgorithmParameters algParameters = AlgorithmParameters
-                    .getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
-            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
-            cp.init(Cipher.DECRYPT_MODE, key, algParameters);
-            fail("Expected NullPointerException or InvalidAlgorithmParameterException");
-        } catch (NullPointerException npe) {
-        } catch (InvalidAlgorithmParameterException e) {
-        }
-
-        try {
-            IvParameterSpec ivSpec = new IvParameterSpec(NONCE_11_BYTE);
-            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
-            cp.init(Cipher.ENCRYPT_MODE, key, ivSpec);
-            fail("Expected NullPointerException or InvalidAlgorithmParameterException");
-        } catch (NullPointerException npe) {
-        } catch (InvalidAlgorithmParameterException e) {
-        }
-
-        try {
-            IvParameterSpec ivSpec = new IvParameterSpec(NONCE_13_BYTE);
-            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
-            cp.init(Cipher.DECRYPT_MODE, key, ivSpec);
-            fail("Expected NullPointerException or InvalidAlgorithmParameterException");
-        } catch (NullPointerException npe) {
-        } catch (InvalidAlgorithmParameterException e) {
-        }
-    }
-
-    @Test
-    public void testChaCha20Poly1305EncryptUpdateAndDoFinalArguments() throws Exception {
-
-        try {
-            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
-            cp.init(Cipher.ENCRYPT_MODE, key, CHACHA20_POLY1305_PARAM_SPEC);
-            cp.update(PLAIN_TEXT);
-            // Update AAD must be done before cipher update, or do final...
-            cp.updateAAD(CHACHA20_POLY1305_AAD, 0, CHACHA20_POLY1305_AAD.length);
-            fail("Did not get expected IllegalStateException(updateAAD)");
-        } catch (Exception e) {
-        }
-
-        try {
-            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
-            cp.init(Cipher.ENCRYPT_MODE, key, CHACHA20_POLY1305_PARAM_SPEC);
             cp.doFinal(PLAIN_TEXT);
-            // Update AAD must be done before cipher update, or do final...
-            cp.updateAAD(CHACHA20_POLY1305_AAD, 0, CHACHA20_POLY1305_AAD.length);
-            fail("Did not get expected IllegalStateException(updateAAD)");
+
         } catch (Exception e) {
+            fail("Got unexpected exception on encrypt/decrypt...");
+        }
+    }
+
+    @Test
+    public void testChaCha20NullParamSpec() throws Exception {
+
+        try {
+            // For ChaCha20 you need to pass in the nonce/iv because
+            // there is no way to get the iv out from this cipher.
+            ChaCha20ParameterSpec chaCha20pec = null;
+            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
+            cp.init(Cipher.ENCRYPT_MODE, key, chaCha20pec);
+            cp.doFinal(PLAIN_TEXT);
+        } catch (Exception e) {
+            fail("Got unexpected exception on encrypt/decrypt...");
+        }
+    }
+
+    @Test
+    public void testChaCha20InvalidParamSpec() throws Exception {
+
+        try {
+            byte[] iv = null;
+            IvParameterSpec ivSpec = new IvParameterSpec(iv);
+            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
+            cp.init(Cipher.ENCRYPT_MODE, key, ivSpec);
+            fail("Expected NullPointerException or InvalidAlgorithmParameterException");
+        } catch (NullPointerException npe) {
+        } catch (InvalidAlgorithmParameterException e) {
         }
 
         try {
-            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
-            cp.updateAAD(CHACHA20_POLY1305_AAD, 0, CHACHA20_POLY1305_AAD.length);
-            fail("Did not get expected IllegalStateException(Cipher not initialized)");
-        } catch (Exception e) {
+            byte[] iv = null;
+            IvParameterSpec ivSpec = new IvParameterSpec(iv);
+            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
+            cp.init(Cipher.DECRYPT_MODE, key, ivSpec);
+            fail("Expected NullPointerException or InvalidAlgorithmParameterException");
+        } catch (NullPointerException npe) {
+        } catch (InvalidAlgorithmParameterException e) {
         }
 
         try {
-            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
+            ChaCha20ParameterSpec chaCha20pec = new ChaCha20ParameterSpec(NONCE_11_BYTE, COUNTER_1);
+            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
+            cp.init(Cipher.ENCRYPT_MODE, key, chaCha20pec);
+            fail("Expected NullPointerException or InvalidAlgorithmParameterException");
+        } catch (NullPointerException npe) {
+        } catch (IllegalArgumentException e) {
+        }
+
+        try {
+            ChaCha20ParameterSpec chaCha20pec = new ChaCha20ParameterSpec(NONCE_13_BYTE, COUNTER_1);
+            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
+            cp.init(Cipher.DECRYPT_MODE, key, chaCha20pec);
+            fail("Expected NullPointerException or InvalidAlgorithmParameterException");
+        } catch (NullPointerException npe) {
+        } catch (IllegalArgumentException e) {
+        }
+    }
+
+    @Test
+    public void testChaCha20EncryptUpdateAndDoFinalArguments() throws Exception {
+
+        try {
+            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
             cp.doFinal(new byte[0]);
             fail("Did not get expected IllegalStateException(Cipher not initialized)");
         } catch (Exception e) {
         }
 
         try {
-            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
+            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
             cp.init(Cipher.ENCRYPT_MODE, key);
             cp.doFinal(null);
             fail("Did not get expected IllegalArgumentException on doFinal(null)");
@@ -414,7 +347,7 @@ public class BaseTestChaCha20Poly1305 extends BaseTestCipher implements ChaCha20
         }
 
         try {
-            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
+            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
             cp.init(Cipher.ENCRYPT_MODE, key);
             cp.doFinal(new byte[0]);
         } catch (Exception e) {
@@ -422,7 +355,7 @@ public class BaseTestChaCha20Poly1305 extends BaseTestCipher implements ChaCha20
         }
 
         try {
-            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
+            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
             cp.init(Cipher.ENCRYPT_MODE, key);
             cp.doFinal(null, 0);
             fail("Did not get expected IllegalArgumentException on doFinal(null, 0)");
@@ -430,7 +363,7 @@ public class BaseTestChaCha20Poly1305 extends BaseTestCipher implements ChaCha20
         }
 
         try {
-            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
+            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
             cp.init(Cipher.ENCRYPT_MODE, key);
             cp.doFinal(null, 1);
             fail("Did not get expected IllegalArgumentException on doFinal(null, 1)");
@@ -438,15 +371,15 @@ public class BaseTestChaCha20Poly1305 extends BaseTestCipher implements ChaCha20
         }
 
         try {
-            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
+            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
             cp.init(Cipher.ENCRYPT_MODE, key);
             cp.doFinal(new byte[0], 0);
-            fail("Did not get expected exception on doFinal(new byte[0], 0)");
         } catch (ShortBufferException e) {
+            fail("Did not expect exception on doFinal(new byte[0], 0)");
         }
 
         try {
-            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
+            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
             cp.init(Cipher.ENCRYPT_MODE, key);
             cp.doFinal(new byte[0], 1);
             fail("Did not get expected IllegalArgumentException on doFinal(new byte[0], 1)");
@@ -454,7 +387,7 @@ public class BaseTestChaCha20Poly1305 extends BaseTestCipher implements ChaCha20
         }
 
         try {
-            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
+            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
             cp.init(Cipher.ENCRYPT_MODE, key);
             cp.doFinal(new byte[cp.getOutputSize(0)], 1);
             fail("Did not get expected IllegalArgumentException on doFinal(new byte[cp.getOutputSize(0)], 1)");
@@ -462,7 +395,7 @@ public class BaseTestChaCha20Poly1305 extends BaseTestCipher implements ChaCha20
         }
 
         try {
-            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
+            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
             cp.init(Cipher.ENCRYPT_MODE, key);
             cp.doFinal(null, 0, 0);
             fail("Did not get expected IllegalArgumentException on doFinal(null, 0, 0)");
@@ -470,7 +403,7 @@ public class BaseTestChaCha20Poly1305 extends BaseTestCipher implements ChaCha20
         }
 
         try {
-            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
+            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
             cp.init(Cipher.ENCRYPT_MODE, key);
             cp.doFinal(null, 1, 0);
             fail("Did not get expected IllegalArgumentException on doFinal(null, 1, 0)");
@@ -478,7 +411,7 @@ public class BaseTestChaCha20Poly1305 extends BaseTestCipher implements ChaCha20
         }
 
         try {
-            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
+            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
             cp.init(Cipher.ENCRYPT_MODE, key);
             cp.doFinal(null, 0, 1);
             fail("Did not get expected IllegalArgumentException on doFinal(null, 0, 1)");
@@ -486,7 +419,7 @@ public class BaseTestChaCha20Poly1305 extends BaseTestCipher implements ChaCha20
         }
 
         try {
-            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
+            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
             cp.init(Cipher.ENCRYPT_MODE, key);
             cp.doFinal(new byte[0], 0, 0);
         } catch (Exception e) {
@@ -494,7 +427,7 @@ public class BaseTestChaCha20Poly1305 extends BaseTestCipher implements ChaCha20
         }
 
         try {
-            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
+            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
             cp.init(Cipher.ENCRYPT_MODE, key);
             cp.doFinal(new byte[0], 1, 0);
             fail("Did not get expected IllegalArgumentException on doFinal(new byte[0], 1, 0)");
@@ -502,7 +435,7 @@ public class BaseTestChaCha20Poly1305 extends BaseTestCipher implements ChaCha20
         }
 
         try {
-            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
+            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
             cp.init(Cipher.ENCRYPT_MODE, key);
             cp.doFinal(new byte[0], 0, 1);
             fail("Did not get expected IllegalArgumentException on doFinal(new byte[0], 0, 1)");
@@ -510,7 +443,7 @@ public class BaseTestChaCha20Poly1305 extends BaseTestCipher implements ChaCha20
         }
 
         try {
-            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
+            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
             cp.init(Cipher.ENCRYPT_MODE, key);
             cp.doFinal(null, 0, 0, null);
             fail("Did not get expected IllegalArgumentException on doFinal(new byte[0], 0, 1)");
@@ -518,15 +451,15 @@ public class BaseTestChaCha20Poly1305 extends BaseTestCipher implements ChaCha20
         }
 
         try {
-            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
+            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
             cp.init(Cipher.ENCRYPT_MODE, key);
             cp.doFinal(new byte[0], 0, 0, new byte[0]);
-            fail("Did not get expected exception on doFinal(new byte[0], 0, 0, new byte[0])");
         } catch (Exception e) {
+            fail("Got unexpected exception on doFinal(new byte[0], 0, 0, new byte[0])");
         }
 
         try {
-            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
+            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
             cp.init(Cipher.ENCRYPT_MODE, key);
             cp.doFinal(new byte[0], 0, 0, null, 0);
             fail("Did not get expected IllegalArgumentException on doFinal(new byte[0], 0, 0, null, 0)");
@@ -534,22 +467,22 @@ public class BaseTestChaCha20Poly1305 extends BaseTestCipher implements ChaCha20
         }
 
         try {
-            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
+            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
             cp.init(Cipher.ENCRYPT_MODE, key);
             cp.doFinal(new byte[0], 0, 0, new byte[0], 0);
-            fail("Did not get expected exception on doFinal(new byte[0], 0, 0, new byte[0], 0)");
         } catch (Exception e) {
+            fail("Got unexpected exception on doFinal(new byte[0], 0, 0, new byte[0], 0)");
         }
 
         try {
-            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
+            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
             cp.update(new byte[0]);
             fail("Did not get expected IllegalStateException(Cipher not initialized)");
         } catch (Exception e) {
         }
 
         try {
-            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
+            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
             cp.init(Cipher.ENCRYPT_MODE, key);
             cp.update(null);
             fail("Did not get expected IllegalArgumentException on update(null)");
@@ -557,7 +490,7 @@ public class BaseTestChaCha20Poly1305 extends BaseTestCipher implements ChaCha20
         }
 
         try {
-            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
+            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
             cp.init(Cipher.ENCRYPT_MODE, key);
             cp.update(new byte[0]);
         } catch (Exception e) {
@@ -565,7 +498,7 @@ public class BaseTestChaCha20Poly1305 extends BaseTestCipher implements ChaCha20
         }
 
         try {
-            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
+            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
             cp.init(Cipher.ENCRYPT_MODE, key);
             cp.update(null, 0, 0);
             fail("Did not get expected IllegalArgumentException on update(null, 0, 0)");
@@ -573,7 +506,7 @@ public class BaseTestChaCha20Poly1305 extends BaseTestCipher implements ChaCha20
         }
 
         try {
-            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
+            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
             cp.init(Cipher.ENCRYPT_MODE, key);
             cp.update(null, 1, 0);
             fail("Did not get expected IllegalArgumentException on update(null, 1, 0)");
@@ -581,7 +514,7 @@ public class BaseTestChaCha20Poly1305 extends BaseTestCipher implements ChaCha20
         }
 
         try {
-            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
+            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
             cp.init(Cipher.ENCRYPT_MODE, key);
             cp.update(null, 0, 1);
             fail("Did not get expected IllegalArgumentException on update(null, 0, 1)");
@@ -589,7 +522,7 @@ public class BaseTestChaCha20Poly1305 extends BaseTestCipher implements ChaCha20
         }
 
         try {
-            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
+            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
             cp.init(Cipher.ENCRYPT_MODE, key);
             cp.update(new byte[0], 0, 0);
         } catch (Exception e) {
@@ -597,7 +530,7 @@ public class BaseTestChaCha20Poly1305 extends BaseTestCipher implements ChaCha20
         }
 
         try {
-            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
+            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
             cp.init(Cipher.ENCRYPT_MODE, key);
             cp.update(new byte[0], 1, 0);
             fail("Did not get expected IllegalArgumentException on update(new byte[0], 1, 0)");
@@ -605,7 +538,7 @@ public class BaseTestChaCha20Poly1305 extends BaseTestCipher implements ChaCha20
         }
 
         try {
-            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
+            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
             cp.init(Cipher.ENCRYPT_MODE, key);
             cp.update(new byte[0], 0, 1);
             fail("Did not get expected IllegalArgumentException on update(new byte[0], 0, 1)");
@@ -613,7 +546,7 @@ public class BaseTestChaCha20Poly1305 extends BaseTestCipher implements ChaCha20
         }
 
         try {
-            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
+            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
             cp.init(Cipher.ENCRYPT_MODE, key);
             cp.update(null, 0, 0, null);
             fail("Did not get expected IllegalArgumentException on update(null, 0, 0, null)");
@@ -621,7 +554,7 @@ public class BaseTestChaCha20Poly1305 extends BaseTestCipher implements ChaCha20
         }
 
         try {
-            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
+            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
             cp.init(Cipher.ENCRYPT_MODE, key);
             cp.update(null, 0, 0, new byte[0]);
             fail("Did not get expected IllegalArgumentException on update(null, 0, 0, new byte[0])");
@@ -629,7 +562,7 @@ public class BaseTestChaCha20Poly1305 extends BaseTestCipher implements ChaCha20
         }
 
         try {
-            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
+            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
             cp.init(Cipher.ENCRYPT_MODE, key);
             cp.update(new byte[0], 0, 0, null);
         } catch (Exception e) {
@@ -637,7 +570,7 @@ public class BaseTestChaCha20Poly1305 extends BaseTestCipher implements ChaCha20
         }
 
         try {
-            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
+            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
             cp.init(Cipher.ENCRYPT_MODE, key);
             cp.update(new byte[0], 0, 0, new byte[0]);
         } catch (Exception e) {
@@ -645,7 +578,7 @@ public class BaseTestChaCha20Poly1305 extends BaseTestCipher implements ChaCha20
         }
 
         try {
-            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
+            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
             cp.init(Cipher.ENCRYPT_MODE, key);
             cp.update(new byte[0], 0, 0, null, 0);
         } catch (Exception e) {
@@ -653,7 +586,7 @@ public class BaseTestChaCha20Poly1305 extends BaseTestCipher implements ChaCha20
         }
 
         try {
-            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
+            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
             cp.init(Cipher.ENCRYPT_MODE, key);
             cp.update(new byte[0], 0, 0, null, 1);
         } catch (Exception e) {
@@ -661,7 +594,7 @@ public class BaseTestChaCha20Poly1305 extends BaseTestCipher implements ChaCha20
         }
 
         try {
-            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
+            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
             cp.init(Cipher.ENCRYPT_MODE, key);
             cp.update(new byte[0], 0, 0, new byte[0], 0);
         } catch (Exception e) {
@@ -669,62 +602,63 @@ public class BaseTestChaCha20Poly1305 extends BaseTestCipher implements ChaCha20
         }
     }
 
+    //--------------------------------------------------------------------------
+    // Run encrypt/decrypt test using MIN/MAX counter values
+    //
     @Test
-    public void testChaCha20Poly1305DecryptUpdateAndDoFinalArguments() throws Exception {
+    public void testChaCha20EncryptDecryptMinMaxCounter() throws Exception {
 
         try {
-            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
-            cp.init(Cipher.ENCRYPT_MODE, key, CHACHA20_POLY1305_PARAM_SPEC);
-            cp.updateAAD(CHACHA20_POLY1305_AAD, 0, CHACHA20_POLY1305_AAD.length);
-            cp.doFinal(PLAIN_TEXT);
+            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
+            cp.init(Cipher.ENCRYPT_MODE, key, CHACHA20_PARAM_SPEC_COUNTER_MIN);
+            byte[] cipherText = cp.doFinal(PLAIN_TEXT);
 
-            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
-            cp.init(Cipher.DECRYPT_MODE, key, CHACHA20_POLY1305_PARAM_SPEC);
-            cp.update(PLAIN_TEXT);
-            // Update AAD must be done before cipher update, or do final...
-            cp.updateAAD(CHACHA20_POLY1305_AAD, 0, CHACHA20_POLY1305_AAD.length);
-            fail("Did not get expected IllegalStateException(updateAAD)");
+            paramSpec = CHACHA20_PARAM_SPEC_COUNTER_MIN;
+
+            // Verify the text
+            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
+            cp.init(Cipher.DECRYPT_MODE, key, paramSpec);
+            byte[] newPlainText = cp.doFinal(cipherText, 0, cipherText.length);
+
+            assertTrue(Arrays.equals(PLAIN_TEXT, newPlainText));
+
         } catch (Exception e) {
+            fail("Got unexpected exception on MIN counter value...");
         }
 
         try {
-            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
-            cp.init(Cipher.ENCRYPT_MODE, key, CHACHA20_POLY1305_PARAM_SPEC);
-            cp.updateAAD(CHACHA20_POLY1305_AAD, 0, CHACHA20_POLY1305_AAD.length);
-            cp.doFinal(PLAIN_TEXT);
+            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
+            cp.init(Cipher.ENCRYPT_MODE, key, CHACHA20_PARAM_SPEC_COUNTER_MAX);
+            byte[] cipherText = cp.doFinal(PLAIN_TEXT);
 
-            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
-            cp.init(Cipher.DECRYPT_MODE, key, CHACHA20_POLY1305_PARAM_SPEC);
-            cp.doFinal(PLAIN_TEXT);
-            // Update AAD must be done before cipher update, or do final...
-            cp.updateAAD(CHACHA20_POLY1305_AAD, 0, CHACHA20_POLY1305_AAD.length);
-            fail("Did not get expected ProviderException(Failure in engineDoFinal)");
+            paramSpec = CHACHA20_PARAM_SPEC_COUNTER_MAX;
+
+            // Verify the text
+            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
+            cp.init(Cipher.DECRYPT_MODE, key, paramSpec);
+            byte[] newPlainText = cp.doFinal(cipherText, 0, cipherText.length);
+
+            assertTrue(Arrays.equals(PLAIN_TEXT, newPlainText));
+
         } catch (Exception e) {
+            fail("Got unexpected exception on MAX counter value...");
         }
-
-        try {
-            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
-            cp.updateAAD(CHACHA20_POLY1305_AAD, 0, CHACHA20_POLY1305_AAD.length);
-            fail("Did not get expected IllegalStateException(Cipher not initialized)");
-        } catch (Exception e) {
-        }
-
     }
 
     //--------------------------------------------------------------------------
     // Run encrypt/decrypt test using just doFinal calls
     //
     @Test
-    public void testChaCha20Poly1305EncryptDecryptDoFinalWithoutAAD() throws Exception {
+    public void testChaCha20EncryptDecryptDoFinal() throws Exception {
         try {
-            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
-            cp.init(Cipher.ENCRYPT_MODE, key, CHACHA20_POLY1305_PARAM_SPEC);
+            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
+            cp.init(Cipher.ENCRYPT_MODE, key, CHACHA20_PARAM_SPEC_COUNTER_0);
             byte[] cipherText = cp.doFinal(PLAIN_TEXT);
 
-            paramSpec = cp.getParameters().getParameterSpec(IvParameterSpec.class);
+            paramSpec = CHACHA20_PARAM_SPEC_COUNTER_0;
 
             // Verify the text
-            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
+            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
             cp.init(Cipher.DECRYPT_MODE, key, paramSpec);
             byte[] newPlainText = cp.doFinal(cipherText, 0, cipherText.length);
 
@@ -732,82 +666,6 @@ public class BaseTestChaCha20Poly1305 extends BaseTestCipher implements ChaCha20
 
         } catch (Exception e) {
             fail("Got unexpected exception on encrypt/decrypt...");
-        }
-    }
-
-    //--------------------------------------------------------------------------
-    // Run encrypt/decrypt test using just doFinal calls
-    //
-    @Test
-    public void testChaCha20Poly1305EncryptDecryptDoFinalWithAAD() throws Exception {
-        try {
-            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
-            cp.init(Cipher.ENCRYPT_MODE, key, CHACHA20_POLY1305_PARAM_SPEC);
-            cp.updateAAD(CHACHA20_POLY1305_AAD, 0, CHACHA20_POLY1305_AAD.length);
-            byte[] cipherText = cp.doFinal(PLAIN_TEXT);
-
-            paramSpec = cp.getParameters().getParameterSpec(IvParameterSpec.class);
-
-            // Verify the text
-            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
-            cp.init(Cipher.DECRYPT_MODE, key, paramSpec);
-            cp.updateAAD(CHACHA20_POLY1305_AAD, 0, CHACHA20_POLY1305_AAD.length);
-            byte[] newPlainText = cp.doFinal(cipherText, 0, cipherText.length);
-
-            assertTrue(Arrays.equals(PLAIN_TEXT, newPlainText));
-
-        } catch (Exception e) {
-            fail("Got unexpected exception on encrypt/decrypt...");
-        }
-    }
-
-    //--------------------------------------------------------------------------
-    // Run encrypt/decrypt test using just doFinal calls
-    //
-    @Test
-    public void testChaCha20Poly1305EncryptDecryptDoFinalWithoutAadBadTag() throws Exception {
-        try {
-            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
-            cp.init(Cipher.ENCRYPT_MODE, key, CHACHA20_POLY1305_PARAM_SPEC);
-            byte[] cipherText = cp.doFinal(PLAIN_TEXT);
-            System.arraycopy(BAD_TAG_16, 0, cipherText, cipherText.length - BAD_TAG_16.length,
-                    BAD_TAG_16.length);
-
-            paramSpec = cp.getParameters().getParameterSpec(IvParameterSpec.class);
-
-            // Verify the text
-            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
-            cp.init(Cipher.DECRYPT_MODE, key, paramSpec);
-            cp.doFinal(cipherText, 0, cipherText.length);
-
-            fail("Did not get expected ProviderException: Failure on doFinal(Bad Tag)");
-
-        } catch (Exception e) {
-        }
-    }
-
-    //--------------------------------------------------------------------------
-    // Run encrypt/decrypt test using just doFinal calls
-    //
-    @Test
-    public void testChaCha20Poly1305DecryptDoFinalWithAadNoTag() throws Exception {
-        try {
-
-            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
-            cp.init(Cipher.ENCRYPT_MODE, key, CHACHA20_POLY1305_PARAM_SPEC);
-            cp.updateAAD(CHACHA20_POLY1305_AAD, 0, CHACHA20_POLY1305_AAD.length);
-            cp.doFinal(PLAIN_TEXT);
-
-            paramSpec = cp.getParameters().getParameterSpec(IvParameterSpec.class);
-
-            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
-            cp.init(Cipher.DECRYPT_MODE, key, paramSpec);
-            cp.updateAAD(CHACHA20_POLY1305_AAD, 0, CHACHA20_POLY1305_AAD.length);
-            cp.doFinal();
-
-            fail("Did not get expected IllegalArgumentException(Missing tag on decrypt final)");
-
-        } catch (Exception e) {
         }
     }
 
@@ -815,17 +673,17 @@ public class BaseTestChaCha20Poly1305 extends BaseTestCipher implements ChaCha20
     // Run encrypt/decrypt test using just update, empty doFinal calls
     //
     @Test
-    public void testChaCha20Poly1305EncryptDecryptUpdate() throws Exception {
+    public void testChaCha20EncryptDecryptUpdate() throws Exception {
         try {
-            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
-            cp.init(Cipher.ENCRYPT_MODE, key, CHACHA20_POLY1305_PARAM_SPEC);
+            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
+            cp.init(Cipher.ENCRYPT_MODE, key, CHACHA20_PARAM_SPEC_COUNTER_1);
             byte[] cipherText1 = cp.update(PLAIN_TEXT);
             byte[] cipherText2 = cp.doFinal();
 
-            paramSpec = cp.getParameters().getParameterSpec(IvParameterSpec.class);
+            paramSpec = CHACHA20_PARAM_SPEC_COUNTER_1;
 
             // Verify the text
-            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
+            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
             cp.init(Cipher.DECRYPT_MODE, key, paramSpec);
             byte[] newPlainText1 = (cipherText1 == null) ? new byte[0] : cp.update(cipherText1);
             byte[] newPlainText2 = cp.doFinal(cipherText2);
@@ -849,19 +707,19 @@ public class BaseTestChaCha20Poly1305 extends BaseTestCipher implements ChaCha20
     // Run encrypt/decrypt test with partial update
     //
     @Test
-    public void testChaCha20Poly1305EncryptDecryptPartialUpdate() throws Exception {
+    public void testChaCha20EncryptDecryptPartialUpdate() throws Exception {
         int partialLen = PLAIN_TEXT.length > 10 ? 10 : 1;
 
         try {
-            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
-            cp.init(Cipher.ENCRYPT_MODE, key, CHACHA20_POLY1305_PARAM_SPEC);
+            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
+            cp.init(Cipher.ENCRYPT_MODE, key, CHACHA20_PARAM_SPEC_COUNTER_0);
             byte[] cipherText1 = cp.update(PLAIN_TEXT, 0, partialLen);
             byte[] cipherText2 = cp.doFinal(PLAIN_TEXT, partialLen, PLAIN_TEXT.length - partialLen);
 
-            paramSpec = cp.getParameters().getParameterSpec(IvParameterSpec.class);
+            paramSpec = CHACHA20_PARAM_SPEC_COUNTER_0;
 
             // Verify the text
-            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
+            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
             cp.init(Cipher.DECRYPT_MODE, key, paramSpec);
             byte[] newPlainText1 = (cipherText1 == null) ? new byte[0] : cp.update(cipherText1);
             byte[] newPlainText2 = cp.doFinal(cipherText2);
@@ -885,21 +743,19 @@ public class BaseTestChaCha20Poly1305 extends BaseTestCipher implements ChaCha20
     // Run encrypt twice using same Cipher instance, same key, but different nonce
     //
     @Test
-    public void testChaCha20Poly1305ReuseObjectSameKeyDifferentNonce() throws Exception {
+    public void testChaCha20ReuseObjectSameKeyDifferentNonce() throws Exception {
         try {
-            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
-            cp.init(Cipher.ENCRYPT_MODE, key, CHACHA20_POLY1305_PARAM_SPEC);
+            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
+            cp.init(Cipher.ENCRYPT_MODE, key, CHACHA20_PARAM_SPEC_COUNTER_0);
             byte[] cipherText1 = cp.doFinal(PLAIN_TEXT);
 
-            paramSpec = cp.getParameters().getParameterSpec(IvParameterSpec.class);
-
+            paramSpec = CHACHA20_PARAM_SPEC_COUNTER_0;
             byte[] newNonce = Arrays.copyOf(NONCE_12_BYTE, NONCE_12_BYTE.length);
             newNonce[0]++;
-            IvParameterSpec newNonceSpec = new IvParameterSpec(newNonce);
+            ChaCha20ParameterSpec newNonceSpec = new ChaCha20ParameterSpec(newNonce, COUNTER_0);
             // Verify that the cipher object can be used to encrypt again with same key different nonce
             cp.init(Cipher.ENCRYPT_MODE, key, newNonceSpec);
             byte[] cipherText2 = cp.doFinal(PLAIN_TEXT);
-
             boolean sameCipher = Arrays.equals(cipherText2, cipherText1);
             assertFalse(sameCipher, "Re-encrypted text with diffent nonce is same");
         } catch (Exception e) {
@@ -911,26 +767,26 @@ public class BaseTestChaCha20Poly1305 extends BaseTestCipher implements ChaCha20
     // Run encrypt/decrypt test using just doFinal calls (copy-safe)
     //
     @Test
-    public void testChaCha20Poly1305EncryptDecryptDoFinalCopySafe() throws Exception {
+    public void testChaCha20EncryptDecryptDoFinalCopySafe() throws Exception {
         try {
-            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
-            cp.init(Cipher.ENCRYPT_MODE, key, CHACHA20_POLY1305_PARAM_SPEC);
+            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
+            cp.init(Cipher.ENCRYPT_MODE, key, CHACHA20_PARAM_SPEC_COUNTER_0);
 
             byte[] cipherText0 = cp.doFinal(PLAIN_TEXT);
 
             byte[] resultBuffer = Arrays.copyOf(PLAIN_TEXT, cp.getOutputSize(PLAIN_TEXT.length));
-            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
-            cp.init(Cipher.ENCRYPT_MODE, key, CHACHA20_POLY1305_PARAM_SPEC);
+            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
+            cp.init(Cipher.ENCRYPT_MODE, key, CHACHA20_PARAM_SPEC_COUNTER_0);
             int resultLen = cp.doFinal(resultBuffer, 0, PLAIN_TEXT.length, resultBuffer);
             byte[] cipherText = Arrays.copyOf(resultBuffer, resultLen);
 
-            paramSpec = cp.getParameters().getParameterSpec(IvParameterSpec.class);;
+            paramSpec = CHACHA20_PARAM_SPEC_COUNTER_0;
 
             boolean success = Arrays.equals(cipherText, cipherText0);
             assertTrue(success, "Encrypted text does not match expected result");
 
             // Verify the text
-            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
+            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
             cp.init(Cipher.DECRYPT_MODE, key, paramSpec);
             resultBuffer = Arrays.copyOf(cipherText, cipherText.length); //cp.getOutputSize(cipherText.length));
             resultLen = cp.doFinal(resultBuffer, 0, cipherText.length, resultBuffer);
@@ -948,16 +804,16 @@ public class BaseTestChaCha20Poly1305 extends BaseTestCipher implements ChaCha20
     // Run encrypt/decrypt test using just update, empty doFinal calls (copy-safe)
     //
     @Test
-    public void testChaCha20Poly1305EncryptDecryptUpdateCopySafe() throws Exception {
+    public void testChaCha20EncryptDecryptUpdateCopySafe() throws Exception {
         try {
-            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
-            cp.init(Cipher.ENCRYPT_MODE, key, CHACHA20_POLY1305_PARAM_SPEC);
+            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
+            cp.init(Cipher.ENCRYPT_MODE, key, CHACHA20_PARAM_SPEC_COUNTER_0);
 
             byte[] cipherText0 = cp.doFinal(PLAIN_TEXT);
 
             byte[] cipherText1 = Arrays.copyOf(PLAIN_TEXT, cp.getOutputSize(PLAIN_TEXT.length));
-            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
-            cp.init(Cipher.ENCRYPT_MODE, key, CHACHA20_POLY1305_PARAM_SPEC);
+            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
+            cp.init(Cipher.ENCRYPT_MODE, key, CHACHA20_PARAM_SPEC_COUNTER_0);
             int cipherText1Len = cp.update(cipherText1, 0, PLAIN_TEXT.length, cipherText1);
             byte[] cipherText2 = cp.doFinal();
 
@@ -965,18 +821,19 @@ public class BaseTestChaCha20Poly1305 extends BaseTestCipher implements ChaCha20
             System.arraycopy(cipherText1, 0, cipherText12, 0, cipherText1Len);
             System.arraycopy(cipherText2, 0, cipherText12, cipherText1Len, cipherText2.length);
 
-            paramSpec = cp.getParameters().getParameterSpec(IvParameterSpec.class);
+            // You can not get the parameters for ChaCha20
+            paramSpec = CHACHA20_PARAM_SPEC_COUNTER_0;
 
             boolean success = Arrays.equals(cipherText12, cipherText0);
             assertTrue(success, "Encrypted text does not match expected result");
 
             // Verify the text
-            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
+            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
             cp.init(Cipher.DECRYPT_MODE, key, paramSpec);
 
             byte[] plainText1 = Arrays.copyOf(cipherText12, cp.getOutputSize(cipherText12.length));
-            int plainText1Len = cp.update(plainText1, 0, plainText1.length, plainText1);
-            byte[] plainText2 = cp.doFinal(cipherText2);
+            int plainText1Len = cp.update(plainText1, 0, cipherText12.length, plainText1);
+            byte[] plainText2 = cp.doFinal();
 
             byte[] plainText12 = new byte[plainText1Len + plainText2.length];
             System.arraycopy(plainText1, 0, plainText12, 0, plainText1Len);
