+++ src/test/java/ibm/jceplus/junit/tests/TestChaCha20Poly1305.java
@@ -1,14 +1,15 @@
 /*
- * Copyright IBM Corp. 2023, 2024
+ * Copyright IBM Corp. 2023, 2026
  *
  * This code is free software; you can redistribute it and/or modify it
  * under the terms provided by IBM in the LICENSE file that accompanied
  * this code, including the "Classpath" Exception described therein.
  */
 
-package ibm.jceplus.junit.base;
+package ibm.jceplus.junit.tests;
 
 import com.ibm.crypto.plus.provider.ChaCha20Constants;
+import java.security.AlgorithmParameters;
 import java.security.InvalidAlgorithmParameterException;
 import java.security.InvalidKeyException;
 import java.security.NoSuchAlgorithmException;
@@ -17,18 +18,27 @@ import java.util.Arrays;
 import javax.crypto.Cipher;
 import javax.crypto.KeyGenerator;
 import javax.crypto.SecretKey;
-import javax.crypto.SecretKeyFactory;
 import javax.crypto.ShortBufferException;
-import javax.crypto.spec.ChaCha20ParameterSpec;
 import javax.crypto.spec.IvParameterSpec;
 import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.TestInstance;
+import org.junit.jupiter.params.Parameter;
+import org.junit.jupiter.params.ParameterizedClass;
+import org.junit.jupiter.params.provider.MethodSource;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.junit.jupiter.api.Assertions.fail;
 
-public class BaseTestChaCha20 extends BaseTestCipher implements ChaCha20Constants {
+@Tag(Tags.OPENJCEPLUS_NAME)
+@TestInstance(TestInstance.Lifecycle.PER_CLASS)
+@ParameterizedClass
+@MethodSource("ibm.jceplus.junit.tests.TestArguments#getOpenJCEPlusOnly")
+public class TestChaCha20Poly1305 extends BaseTest implements ChaCha20Constants {
 
+    @Parameter(0)
+    TestProvider provider;
 
     // 14 bytes: PASSED
     static final byte[] PLAIN_TEXT_14 = "12345678123456".getBytes();
@@ -54,33 +64,26 @@ public class BaseTestChaCha20 extends BaseTestCipher implements ChaCha20Constant
     static final byte[] NONCEA_12_BYTE = "012345678123".getBytes();
     static final byte[] NONCE_13_BYTE = "1234567812345".getBytes();
 
-    static final int COUNTER_0 = 0;
-    static final int COUNTER_1 = 1;
-    static final int COUNTER_MIN = Integer.MIN_VALUE;
-    static final int COUNTER_MAX = Integer.MAX_VALUE;
+    static final byte[] BAD_TAG_16 = "BaadTaagBaadTaag".getBytes();
 
+    static final byte[] CHACHA20_POLY1305_AAD = "12345".getBytes(); //"ChaCha20-Poly1305 AAD".getBytes();
+
+    static final String CHACHA20_POLY1305_ALGORITHM = "ChaCha20-Poly1305";
     static final String CHACHA20_ALGORITHM = "ChaCha20";
 
-    static final ChaCha20ParameterSpec CHACHA20_PARAM_SPEC_COUNTER_0 = new ChaCha20ParameterSpec(
-            NONCE_12_BYTE, COUNTER_0);
-    static final ChaCha20ParameterSpec CHACHA20_PARAM_SPEC_COUNTER_1 = new ChaCha20ParameterSpec(
-            NONCE_12_BYTE, COUNTER_1);
-    static final ChaCha20ParameterSpec CHACHA20_PARAM_SPEC_COUNTER_MIN = new ChaCha20ParameterSpec(
-            NONCE_12_BYTE, COUNTER_MIN);
-    static final ChaCha20ParameterSpec CHACHA20_PARAM_SPEC_COUNTER_MAX = new ChaCha20ParameterSpec(
-            NONCE_12_BYTE, COUNTER_MAX);
+    static final IvParameterSpec CHACHA20_POLY1305_PARAM_SPEC = new IvParameterSpec(NONCE_12_BYTE);
 
 
     protected KeyGenerator keyGen = null;
     protected SecretKey key = null;
-    protected SecretKeyFactory keyFactory = null;
-    protected ChaCha20ParameterSpec paramSpec = null;
+    protected IvParameterSpec paramSpec = null;
     protected Cipher cp = null;
     protected boolean success = true;
     protected int specifiedKeySize = 0;
 
     @BeforeEach
     public void setUp() throws Exception {
+        setAndInsertProvider(provider);
         keyGen = KeyGenerator.getInstance(CHACHA20_ALGORITHM, getProviderName());
         if (specifiedKeySize > 0) {
             keyGen.init(specifiedKeySize);
@@ -89,33 +92,24 @@ public class BaseTestChaCha20 extends BaseTestCipher implements ChaCha20Constant
     }
 
     @Test
-    public void testChaCha20KeyFactory() throws Exception {
-
+    public void testChaCha20Poly1305IllegalKeyNonceReuse() throws Exception {
         try {
-            keyFactory = SecretKeyFactory.getInstance(CHACHA20_ALGORITHM, getProviderName());
-        } catch (Exception e) {
-            fail("Got unexpected exception on ChaCha20KeyFactory.getInstance()...");
-        }
-    }
-
-    @Test
-    public void testChaCha20IllegalKeyNonceReuse() throws Exception {
-        try {
-            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
-            cp.init(Cipher.ENCRYPT_MODE, key, CHACHA20_PARAM_SPEC_COUNTER_0);
-            cp.init(Cipher.DECRYPT_MODE, key, CHACHA20_PARAM_SPEC_COUNTER_0);
+            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
+            cp.init(Cipher.ENCRYPT_MODE, key, CHACHA20_POLY1305_PARAM_SPEC);
+            cp.init(Cipher.DECRYPT_MODE, key, CHACHA20_POLY1305_PARAM_SPEC);
 
             //fail("Expected InvalidKeyException did not occur");
 
         } catch (InvalidKeyException e) {
-            assertTrue(false);
+            e.printStackTrace();
+            throw e;
         }
     }
 
     @Test
-    public void testChaCha20ShortBuffer() throws Exception {
+    public void testChaCha20Poly1305ShortBuffer() throws Exception {
         try {
-            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
+            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
 
             // Encrypt the plain text
             cp.init(Cipher.ENCRYPT_MODE, key);
@@ -129,13 +123,13 @@ public class BaseTestChaCha20 extends BaseTestCipher implements ChaCha20Constant
 
     @Test
     public void testChaCha20Poly1305EncryptParamsAfterShortBufferRetry() throws Exception {
-        cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
-        ChaCha20ParameterSpec paraSpec = new ChaCha20ParameterSpec(NONCE_12_BYTE, COUNTER_0);
+        cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
+        IvParameterSpec ivParamSpec = new IvParameterSpec(NONCE_12_BYTE);
 
         // Try encrypting with a short buffer.
         try {
-            cp.init(Cipher.ENCRYPT_MODE, key, paraSpec);
-            byte[] cipherText = new byte[PLAIN_TEXT.length - 1];
+            cp.init(Cipher.ENCRYPT_MODE, key, ivParamSpec);
+            byte[] cipherText = new byte[PLAIN_TEXT.length + 15];
             cp.doFinal(PLAIN_TEXT, 0, PLAIN_TEXT.length, cipherText);
             fail("Expected ShortBufferException did not occur");
         } catch (ShortBufferException sbe) {
@@ -152,20 +146,20 @@ public class BaseTestChaCha20 extends BaseTestCipher implements ChaCha20Constant
         }
 
         // Retry with a larger buffer after re-initializing.
-        paraSpec = new ChaCha20ParameterSpec(NONCEA_12_BYTE, COUNTER_0);
-        cp.init(Cipher.ENCRYPT_MODE, key, paraSpec);
+        ivParamSpec = new IvParameterSpec(NONCEA_12_BYTE);
+        cp.init(Cipher.ENCRYPT_MODE, key, ivParamSpec);
         byte[] largerCipherTextBuffer = new byte[PLAIN_TEXT.length + 16];
         cp.doFinal(PLAIN_TEXT, 0, PLAIN_TEXT.length, largerCipherTextBuffer);
     }
 
     @Test
     public void testChaCha20Poly1305EncryptNoParamsAfterShortBufferRetry() throws Exception {
-        cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
+        cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
 
         // Try encrypting with a short buffer.
         try {
             cp.init(Cipher.ENCRYPT_MODE, key);
-            byte[] cipherText = new byte[PLAIN_TEXT.length - 1];
+            byte[] cipherText = new byte[PLAIN_TEXT.length + 15];
             cp.doFinal(PLAIN_TEXT, 0, PLAIN_TEXT.length, cipherText);
             fail("Expected ShortBufferException did not occur");
         } catch (ShortBufferException sbe) {
@@ -188,26 +182,62 @@ public class BaseTestChaCha20 extends BaseTestCipher implements ChaCha20Constant
     }
 
     @Test
-    public void testChaCha20GetBlockSizeEncryptDecrypt() throws Exception {
+    public void testChaCha20Poly1305DecryptAfterShortBufferRetry() throws Exception {
+        byte[] cipherText = null;
+        Cipher cpl = null;
+        try {
+            cpl = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
+            IvParameterSpec ivParamSpec = new IvParameterSpec(NONCEA_12_BYTE);
+
+            // Encrypt the plain text
+            cpl.init(Cipher.ENCRYPT_MODE, key, ivParamSpec);
+            cipherText = cpl.doFinal(PLAIN_TEXT, 0, PLAIN_TEXT.length);
+
+            ivParamSpec = cpl.getParameters().getParameterSpec(IvParameterSpec.class);
+
+            cpl = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
+            cpl.init(Cipher.DECRYPT_MODE, key, ivParamSpec);
+            byte[] sbPlainText = new byte[15];
+            System.out.println("cipherText.length=" + cipherText.length);
+            System.out.println("sbPlainText.length=" + sbPlainText.length);
+            cpl.doFinal(cipherText, 0, cipherText.length, sbPlainText, 0);
+            fail("Failed to get ShortedBufferException");
+
+        } catch (ShortBufferException ex) {
+            System.out.println("Try retry with a larger buffer");
+        }
+        // try retry with a larger buffer
+        try {
+            byte[] lbPlainTextBuffer = new byte[PLAIN_TEXT.length];
+            cpl.doFinal(cipherText, 0, cipherText.length, lbPlainTextBuffer, 0);
+            assertTrue(Arrays.equals(PLAIN_TEXT, lbPlainTextBuffer));
+        } catch (Exception ex) {
+            ex.printStackTrace();
+            fail("Retying with larger buffer should have worked  with a larger buffer");
+        }
+    }
+
+    @Test
+    public void testChaCha20Poly1305GetBlockSizeEncryptDecrypt() throws Exception {
         chaCha20GetBlockSize(Cipher.ENCRYPT_MODE);
         chaCha20GetBlockSize(Cipher.DECRYPT_MODE);
     }
 
 
     public void chaCha20GetBlockSize(int opMode) throws Exception {
-        ChaCha20ParameterSpec chaCha20ParamSpec = new ChaCha20ParameterSpec(NONCE_12_BYTE, 0);
-        cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
+        IvParameterSpec chaCha20ParamSpec = new IvParameterSpec(NONCE_12_BYTE);
+        cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
         cp.init(opMode, key, chaCha20ParamSpec);
         assertTrue((cp.getBlockSize() == ChaCha20_BLOCK_SIZE), "ChaCha20 Block size must be: " + ChaCha20_BLOCK_SIZE);
     }
 
     @Test
-    public void testChaCha20ValidTransformations() throws Exception {
+    public void testChaCha20Poly1305ValidTransformations() throws Exception {
         String transformation = null;
         try {
-            transformation = CHACHA20_ALGORITHM;
+            transformation = CHACHA20_POLY1305_ALGORITHM;
             cp = Cipher.getInstance(transformation, getProviderName());
-            transformation = "ChaCha20/None/NoPadding";
+            transformation = "ChaCha20-Poly1305/None/NoPadding";
             cp = Cipher.getInstance(transformation, getProviderName());
         } catch (NoSuchAlgorithmException ex) {
             fail("NoSuchAlgorithmException occurred for transform: " + transformation);
@@ -215,8 +245,8 @@ public class BaseTestChaCha20 extends BaseTestCipher implements ChaCha20Constant
     }
 
     @Test
-    public void testChaCha20InvalidTransformation() throws Exception {
-        String transformation = "BogusChaCha20/BogusMode/BogusPadding";
+    public void testChaCha20Poly1305InvalidTransformation() throws Exception {
+        String transformation = "BogusChaCha20-Poly1305/BogusMode/BogusPadding";
         try {
             cp = Cipher.getInstance(transformation, getProviderName());
             fail("Expected NoSuchAlgorithmException did not occur");
@@ -226,8 +256,8 @@ public class BaseTestChaCha20 extends BaseTestCipher implements ChaCha20Constant
     }
 
     @Test
-    public void testChaCha20NullKey() throws Exception {
-        cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
+    public void testChaCha20Poly1305NullKey() throws Exception {
+        cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
         SecretKey nullKey = null;
 
         try {
@@ -237,7 +267,6 @@ public class BaseTestChaCha20 extends BaseTestCipher implements ChaCha20Constant
         }
 
         try {
-            //cp.init(Cipher.ENCRYPT_MODE, nullKey, SecureRandom.getInstance("SecureRandom"));
             cp.init(Cipher.ENCRYPT_MODE, nullKey, SecureRandom.getInstance("SHA2DRBG"));
             fail("Expected InvalidKeyException did not occur");
         } catch (InvalidKeyException e) {
@@ -245,14 +274,21 @@ public class BaseTestChaCha20 extends BaseTestCipher implements ChaCha20Constant
     }
 
     @Test
-    public void testChaCha20NoParamSpec() throws Exception {
+    public void testChaCha20Poly1305NoParamSpec() throws Exception {
 
         try {
-            // For ChaCha20 you need to pass in the nonce/iv because
-            // there is no way to get the iv out from this cipher.
-            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
+            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
             cp.init(Cipher.ENCRYPT_MODE, key);
-            cp.doFinal(PLAIN_TEXT);
+            byte[] cipherText = cp.doFinal(PLAIN_TEXT);
+
+            paramSpec = cp.getParameters().getParameterSpec(IvParameterSpec.class);
+
+            // Verify the text
+            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
+            cp.init(Cipher.DECRYPT_MODE, key, paramSpec);
+            byte[] newPlainText = cp.doFinal(cipherText, 0, cipherText.length);
+
+            assertTrue(Arrays.equals(PLAIN_TEXT, newPlainText));
 
         } catch (Exception e) {
             fail("Got unexpected exception on encrypt/decrypt...");
@@ -260,27 +296,35 @@ public class BaseTestChaCha20 extends BaseTestCipher implements ChaCha20Constant
     }
 
     @Test
-    public void testChaCha20NullParamSpec() throws Exception {
+    public void testChaCha20Poly1305NullParamSpec() throws Exception {
 
         try {
-            // For ChaCha20 you need to pass in the nonce/iv because
-            // there is no way to get the iv out from this cipher.
-            ChaCha20ParameterSpec chaCha20pec = null;
-            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
-            cp.init(Cipher.ENCRYPT_MODE, key, chaCha20pec);
-            cp.doFinal(PLAIN_TEXT);
+            IvParameterSpec ivSpec = null;
+            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
+            cp.init(Cipher.ENCRYPT_MODE, key, ivSpec);
+            byte[] cipherText = cp.doFinal(PLAIN_TEXT);
+
+            paramSpec = cp.getParameters().getParameterSpec(IvParameterSpec.class);
+
+            // Verify the text
+            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
+            cp.init(Cipher.DECRYPT_MODE, key, paramSpec);
+            byte[] newPlainText = cp.doFinal(cipherText, 0, cipherText.length);
+
+            assertTrue(Arrays.equals(PLAIN_TEXT, newPlainText));
+
         } catch (Exception e) {
             fail("Got unexpected exception on encrypt/decrypt...");
         }
     }
 
     @Test
-    public void testChaCha20InvalidParamSpec() throws Exception {
+    public void testChaCha20Poly1305InvalidParamSpec() throws Exception {
 
         try {
             byte[] iv = null;
             IvParameterSpec ivSpec = new IvParameterSpec(iv);
-            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
+            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
             cp.init(Cipher.ENCRYPT_MODE, key, ivSpec);
             fail("Expected NullPointerException or InvalidAlgorithmParameterException");
         } catch (NullPointerException npe) {
@@ -290,7 +334,7 @@ public class BaseTestChaCha20 extends BaseTestCipher implements ChaCha20Constant
         try {
             byte[] iv = null;
             IvParameterSpec ivSpec = new IvParameterSpec(iv);
-            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
+            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
             cp.init(Cipher.DECRYPT_MODE, key, ivSpec);
             fail("Expected NullPointerException or InvalidAlgorithmParameterException");
         } catch (NullPointerException npe) {
@@ -298,36 +342,83 @@ public class BaseTestChaCha20 extends BaseTestCipher implements ChaCha20Constant
         }
 
         try {
-            ChaCha20ParameterSpec chaCha20pec = new ChaCha20ParameterSpec(NONCE_11_BYTE, COUNTER_1);
-            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
-            cp.init(Cipher.ENCRYPT_MODE, key, chaCha20pec);
+            AlgorithmParameters algParameters = AlgorithmParameters
+                    .getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
+            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
+            cp.init(Cipher.ENCRYPT_MODE, key, algParameters);
             fail("Expected NullPointerException or InvalidAlgorithmParameterException");
         } catch (NullPointerException npe) {
-        } catch (IllegalArgumentException e) {
+        } catch (InvalidAlgorithmParameterException e) {
         }
 
         try {
-            ChaCha20ParameterSpec chaCha20pec = new ChaCha20ParameterSpec(NONCE_13_BYTE, COUNTER_1);
-            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
-            cp.init(Cipher.DECRYPT_MODE, key, chaCha20pec);
+            AlgorithmParameters algParameters = AlgorithmParameters
+                    .getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
+            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
+            cp.init(Cipher.DECRYPT_MODE, key, algParameters);
             fail("Expected NullPointerException or InvalidAlgorithmParameterException");
         } catch (NullPointerException npe) {
-        } catch (IllegalArgumentException e) {
+        } catch (InvalidAlgorithmParameterException e) {
+        }
+
+        try {
+            IvParameterSpec ivSpec = new IvParameterSpec(NONCE_11_BYTE);
+            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
+            cp.init(Cipher.ENCRYPT_MODE, key, ivSpec);
+            fail("Expected NullPointerException or InvalidAlgorithmParameterException");
+        } catch (NullPointerException npe) {
+        } catch (InvalidAlgorithmParameterException e) {
+        }
+
+        try {
+            IvParameterSpec ivSpec = new IvParameterSpec(NONCE_13_BYTE);
+            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
+            cp.init(Cipher.DECRYPT_MODE, key, ivSpec);
+            fail("Expected NullPointerException or InvalidAlgorithmParameterException");
+        } catch (NullPointerException npe) {
+        } catch (InvalidAlgorithmParameterException e) {
         }
     }
 
     @Test
-    public void testChaCha20EncryptUpdateAndDoFinalArguments() throws Exception {
+    public void testChaCha20Poly1305EncryptUpdateAndDoFinalArguments() throws Exception {
 
         try {
-            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
+            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
+            cp.init(Cipher.ENCRYPT_MODE, key, CHACHA20_POLY1305_PARAM_SPEC);
+            cp.update(PLAIN_TEXT);
+            // Update AAD must be done before cipher update, or do final...
+            cp.updateAAD(CHACHA20_POLY1305_AAD, 0, CHACHA20_POLY1305_AAD.length);
+            fail("Did not get expected IllegalStateException(updateAAD)");
+        } catch (Exception e) {
+        }
+
+        try {
+            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
+            cp.init(Cipher.ENCRYPT_MODE, key, CHACHA20_POLY1305_PARAM_SPEC);
+            cp.doFinal(PLAIN_TEXT);
+            // Update AAD must be done before cipher update, or do final...
+            cp.updateAAD(CHACHA20_POLY1305_AAD, 0, CHACHA20_POLY1305_AAD.length);
+            fail("Did not get expected IllegalStateException(updateAAD)");
+        } catch (Exception e) {
+        }
+
+        try {
+            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
+            cp.updateAAD(CHACHA20_POLY1305_AAD, 0, CHACHA20_POLY1305_AAD.length);
+            fail("Did not get expected IllegalStateException(Cipher not initialized)");
+        } catch (Exception e) {
+        }
+
+        try {
+            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
             cp.doFinal(new byte[0]);
             fail("Did not get expected IllegalStateException(Cipher not initialized)");
         } catch (Exception e) {
         }
 
         try {
-            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
+            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
             cp.init(Cipher.ENCRYPT_MODE, key);
             cp.doFinal(null);
             fail("Did not get expected IllegalArgumentException on doFinal(null)");
@@ -335,7 +426,7 @@ public class BaseTestChaCha20 extends BaseTestCipher implements ChaCha20Constant
         }
 
         try {
-            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
+            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
             cp.init(Cipher.ENCRYPT_MODE, key);
             cp.doFinal(new byte[0]);
         } catch (Exception e) {
@@ -343,7 +434,7 @@ public class BaseTestChaCha20 extends BaseTestCipher implements ChaCha20Constant
         }
 
         try {
-            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
+            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
             cp.init(Cipher.ENCRYPT_MODE, key);
             cp.doFinal(null, 0);
             fail("Did not get expected IllegalArgumentException on doFinal(null, 0)");
@@ -351,7 +442,7 @@ public class BaseTestChaCha20 extends BaseTestCipher implements ChaCha20Constant
         }
 
         try {
-            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
+            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
             cp.init(Cipher.ENCRYPT_MODE, key);
             cp.doFinal(null, 1);
             fail("Did not get expected IllegalArgumentException on doFinal(null, 1)");
@@ -359,15 +450,15 @@ public class BaseTestChaCha20 extends BaseTestCipher implements ChaCha20Constant
         }
 
         try {
-            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
+            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
             cp.init(Cipher.ENCRYPT_MODE, key);
             cp.doFinal(new byte[0], 0);
+            fail("Did not get expected exception on doFinal(new byte[0], 0)");
         } catch (ShortBufferException e) {
-            fail("Did not expect exception on doFinal(new byte[0], 0)");
         }
 
         try {
-            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
+            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
             cp.init(Cipher.ENCRYPT_MODE, key);
             cp.doFinal(new byte[0], 1);
             fail("Did not get expected IllegalArgumentException on doFinal(new byte[0], 1)");
@@ -375,7 +466,7 @@ public class BaseTestChaCha20 extends BaseTestCipher implements ChaCha20Constant
         }
 
         try {
-            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
+            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
             cp.init(Cipher.ENCRYPT_MODE, key);
             cp.doFinal(new byte[cp.getOutputSize(0)], 1);
             fail("Did not get expected IllegalArgumentException on doFinal(new byte[cp.getOutputSize(0)], 1)");
@@ -383,7 +474,7 @@ public class BaseTestChaCha20 extends BaseTestCipher implements ChaCha20Constant
         }
 
         try {
-            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
+            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
             cp.init(Cipher.ENCRYPT_MODE, key);
             cp.doFinal(null, 0, 0);
             fail("Did not get expected IllegalArgumentException on doFinal(null, 0, 0)");
@@ -391,7 +482,7 @@ public class BaseTestChaCha20 extends BaseTestCipher implements ChaCha20Constant
         }
 
         try {
-            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
+            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
             cp.init(Cipher.ENCRYPT_MODE, key);
             cp.doFinal(null, 1, 0);
             fail("Did not get expected IllegalArgumentException on doFinal(null, 1, 0)");
@@ -399,7 +490,7 @@ public class BaseTestChaCha20 extends BaseTestCipher implements ChaCha20Constant
         }
 
         try {
-            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
+            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
             cp.init(Cipher.ENCRYPT_MODE, key);
             cp.doFinal(null, 0, 1);
             fail("Did not get expected IllegalArgumentException on doFinal(null, 0, 1)");
@@ -407,7 +498,7 @@ public class BaseTestChaCha20 extends BaseTestCipher implements ChaCha20Constant
         }
 
         try {
-            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
+            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
             cp.init(Cipher.ENCRYPT_MODE, key);
             cp.doFinal(new byte[0], 0, 0);
         } catch (Exception e) {
@@ -415,7 +506,7 @@ public class BaseTestChaCha20 extends BaseTestCipher implements ChaCha20Constant
         }
 
         try {
-            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
+            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
             cp.init(Cipher.ENCRYPT_MODE, key);
             cp.doFinal(new byte[0], 1, 0);
             fail("Did not get expected IllegalArgumentException on doFinal(new byte[0], 1, 0)");
@@ -423,7 +514,7 @@ public class BaseTestChaCha20 extends BaseTestCipher implements ChaCha20Constant
         }
 
         try {
-            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
+            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
             cp.init(Cipher.ENCRYPT_MODE, key);
             cp.doFinal(new byte[0], 0, 1);
             fail("Did not get expected IllegalArgumentException on doFinal(new byte[0], 0, 1)");
@@ -431,7 +522,7 @@ public class BaseTestChaCha20 extends BaseTestCipher implements ChaCha20Constant
         }
 
         try {
-            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
+            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
             cp.init(Cipher.ENCRYPT_MODE, key);
             cp.doFinal(null, 0, 0, null);
             fail("Did not get expected IllegalArgumentException on doFinal(new byte[0], 0, 1)");
@@ -439,15 +530,15 @@ public class BaseTestChaCha20 extends BaseTestCipher implements ChaCha20Constant
         }
 
         try {
-            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
+            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
             cp.init(Cipher.ENCRYPT_MODE, key);
             cp.doFinal(new byte[0], 0, 0, new byte[0]);
+            fail("Did not get expected exception on doFinal(new byte[0], 0, 0, new byte[0])");
         } catch (Exception e) {
-            fail("Got unexpected exception on doFinal(new byte[0], 0, 0, new byte[0])");
         }
 
         try {
-            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
+            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
             cp.init(Cipher.ENCRYPT_MODE, key);
             cp.doFinal(new byte[0], 0, 0, null, 0);
             fail("Did not get expected IllegalArgumentException on doFinal(new byte[0], 0, 0, null, 0)");
@@ -455,22 +546,22 @@ public class BaseTestChaCha20 extends BaseTestCipher implements ChaCha20Constant
         }
 
         try {
-            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
+            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
             cp.init(Cipher.ENCRYPT_MODE, key);
             cp.doFinal(new byte[0], 0, 0, new byte[0], 0);
+            fail("Did not get expected exception on doFinal(new byte[0], 0, 0, new byte[0], 0)");
         } catch (Exception e) {
-            fail("Got unexpected exception on doFinal(new byte[0], 0, 0, new byte[0], 0)");
         }
 
         try {
-            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
+            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
             cp.update(new byte[0]);
             fail("Did not get expected IllegalStateException(Cipher not initialized)");
         } catch (Exception e) {
         }
 
         try {
-            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
+            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
             cp.init(Cipher.ENCRYPT_MODE, key);
             cp.update(null);
             fail("Did not get expected IllegalArgumentException on update(null)");
@@ -478,7 +569,7 @@ public class BaseTestChaCha20 extends BaseTestCipher implements ChaCha20Constant
         }
 
         try {
-            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
+            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
             cp.init(Cipher.ENCRYPT_MODE, key);
             cp.update(new byte[0]);
         } catch (Exception e) {
@@ -486,7 +577,7 @@ public class BaseTestChaCha20 extends BaseTestCipher implements ChaCha20Constant
         }
 
         try {
-            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
+            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
             cp.init(Cipher.ENCRYPT_MODE, key);
             cp.update(null, 0, 0);
             fail("Did not get expected IllegalArgumentException on update(null, 0, 0)");
@@ -494,7 +585,7 @@ public class BaseTestChaCha20 extends BaseTestCipher implements ChaCha20Constant
         }
 
         try {
-            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
+            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
             cp.init(Cipher.ENCRYPT_MODE, key);
             cp.update(null, 1, 0);
             fail("Did not get expected IllegalArgumentException on update(null, 1, 0)");
@@ -502,7 +593,7 @@ public class BaseTestChaCha20 extends BaseTestCipher implements ChaCha20Constant
         }
 
         try {
-            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
+            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
             cp.init(Cipher.ENCRYPT_MODE, key);
             cp.update(null, 0, 1);
             fail("Did not get expected IllegalArgumentException on update(null, 0, 1)");
@@ -510,7 +601,7 @@ public class BaseTestChaCha20 extends BaseTestCipher implements ChaCha20Constant
         }
 
         try {
-            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
+            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
             cp.init(Cipher.ENCRYPT_MODE, key);
             cp.update(new byte[0], 0, 0);
         } catch (Exception e) {
@@ -518,7 +609,7 @@ public class BaseTestChaCha20 extends BaseTestCipher implements ChaCha20Constant
         }
 
         try {
-            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
+            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
             cp.init(Cipher.ENCRYPT_MODE, key);
             cp.update(new byte[0], 1, 0);
             fail("Did not get expected IllegalArgumentException on update(new byte[0], 1, 0)");
@@ -526,7 +617,7 @@ public class BaseTestChaCha20 extends BaseTestCipher implements ChaCha20Constant
         }
 
         try {
-            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
+            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
             cp.init(Cipher.ENCRYPT_MODE, key);
             cp.update(new byte[0], 0, 1);
             fail("Did not get expected IllegalArgumentException on update(new byte[0], 0, 1)");
@@ -534,7 +625,7 @@ public class BaseTestChaCha20 extends BaseTestCipher implements ChaCha20Constant
         }
 
         try {
-            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
+            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
             cp.init(Cipher.ENCRYPT_MODE, key);
             cp.update(null, 0, 0, null);
             fail("Did not get expected IllegalArgumentException on update(null, 0, 0, null)");
@@ -542,7 +633,7 @@ public class BaseTestChaCha20 extends BaseTestCipher implements ChaCha20Constant
         }
 
         try {
-            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
+            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
             cp.init(Cipher.ENCRYPT_MODE, key);
             cp.update(null, 0, 0, new byte[0]);
             fail("Did not get expected IllegalArgumentException on update(null, 0, 0, new byte[0])");
@@ -550,7 +641,7 @@ public class BaseTestChaCha20 extends BaseTestCipher implements ChaCha20Constant
         }
 
         try {
-            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
+            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
             cp.init(Cipher.ENCRYPT_MODE, key);
             cp.update(new byte[0], 0, 0, null);
         } catch (Exception e) {
@@ -558,7 +649,7 @@ public class BaseTestChaCha20 extends BaseTestCipher implements ChaCha20Constant
         }
 
         try {
-            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
+            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
             cp.init(Cipher.ENCRYPT_MODE, key);
             cp.update(new byte[0], 0, 0, new byte[0]);
         } catch (Exception e) {
@@ -566,7 +657,7 @@ public class BaseTestChaCha20 extends BaseTestCipher implements ChaCha20Constant
         }
 
         try {
-            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
+            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
             cp.init(Cipher.ENCRYPT_MODE, key);
             cp.update(new byte[0], 0, 0, null, 0);
         } catch (Exception e) {
@@ -574,7 +665,7 @@ public class BaseTestChaCha20 extends BaseTestCipher implements ChaCha20Constant
         }
 
         try {
-            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
+            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
             cp.init(Cipher.ENCRYPT_MODE, key);
             cp.update(new byte[0], 0, 0, null, 1);
         } catch (Exception e) {
@@ -582,7 +673,7 @@ public class BaseTestChaCha20 extends BaseTestCipher implements ChaCha20Constant
         }
 
         try {
-            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
+            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
             cp.init(Cipher.ENCRYPT_MODE, key);
             cp.update(new byte[0], 0, 0, new byte[0], 0);
         } catch (Exception e) {
@@ -590,63 +681,62 @@ public class BaseTestChaCha20 extends BaseTestCipher implements ChaCha20Constant
         }
     }
 
-    //--------------------------------------------------------------------------
-    // Run encrypt/decrypt test using MIN/MAX counter values
-    //
     @Test
-    public void testChaCha20EncryptDecryptMinMaxCounter() throws Exception {
+    public void testChaCha20Poly1305DecryptUpdateAndDoFinalArguments() throws Exception {
 
         try {
-            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
-            cp.init(Cipher.ENCRYPT_MODE, key, CHACHA20_PARAM_SPEC_COUNTER_MIN);
-            byte[] cipherText = cp.doFinal(PLAIN_TEXT);
+            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
+            cp.init(Cipher.ENCRYPT_MODE, key, CHACHA20_POLY1305_PARAM_SPEC);
+            cp.updateAAD(CHACHA20_POLY1305_AAD, 0, CHACHA20_POLY1305_AAD.length);
+            cp.doFinal(PLAIN_TEXT);
 
-            paramSpec = CHACHA20_PARAM_SPEC_COUNTER_MIN;
-
-            // Verify the text
-            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
-            cp.init(Cipher.DECRYPT_MODE, key, paramSpec);
-            byte[] newPlainText = cp.doFinal(cipherText, 0, cipherText.length);
-
-            assertTrue(Arrays.equals(PLAIN_TEXT, newPlainText));
-
+            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
+            cp.init(Cipher.DECRYPT_MODE, key, CHACHA20_POLY1305_PARAM_SPEC);
+            cp.update(PLAIN_TEXT);
+            // Update AAD must be done before cipher update, or do final...
+            cp.updateAAD(CHACHA20_POLY1305_AAD, 0, CHACHA20_POLY1305_AAD.length);
+            fail("Did not get expected IllegalStateException(updateAAD)");
         } catch (Exception e) {
-            fail("Got unexpected exception on MIN counter value...");
         }
 
         try {
-            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
-            cp.init(Cipher.ENCRYPT_MODE, key, CHACHA20_PARAM_SPEC_COUNTER_MAX);
-            byte[] cipherText = cp.doFinal(PLAIN_TEXT);
+            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
+            cp.init(Cipher.ENCRYPT_MODE, key, CHACHA20_POLY1305_PARAM_SPEC);
+            cp.updateAAD(CHACHA20_POLY1305_AAD, 0, CHACHA20_POLY1305_AAD.length);
+            cp.doFinal(PLAIN_TEXT);
 
-            paramSpec = CHACHA20_PARAM_SPEC_COUNTER_MAX;
-
-            // Verify the text
-            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
-            cp.init(Cipher.DECRYPT_MODE, key, paramSpec);
-            byte[] newPlainText = cp.doFinal(cipherText, 0, cipherText.length);
-
-            assertTrue(Arrays.equals(PLAIN_TEXT, newPlainText));
-
+            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
+            cp.init(Cipher.DECRYPT_MODE, key, CHACHA20_POLY1305_PARAM_SPEC);
+            cp.doFinal(PLAIN_TEXT);
+            // Update AAD must be done before cipher update, or do final...
+            cp.updateAAD(CHACHA20_POLY1305_AAD, 0, CHACHA20_POLY1305_AAD.length);
+            fail("Did not get expected ProviderException(Failure in engineDoFinal)");
         } catch (Exception e) {
-            fail("Got unexpected exception on MAX counter value...");
         }
+
+        try {
+            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
+            cp.updateAAD(CHACHA20_POLY1305_AAD, 0, CHACHA20_POLY1305_AAD.length);
+            fail("Did not get expected IllegalStateException(Cipher not initialized)");
+        } catch (Exception e) {
+        }
+
     }
 
     //--------------------------------------------------------------------------
     // Run encrypt/decrypt test using just doFinal calls
     //
     @Test
-    public void testChaCha20EncryptDecryptDoFinal() throws Exception {
+    public void testChaCha20Poly1305EncryptDecryptDoFinalWithoutAAD() throws Exception {
         try {
-            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
-            cp.init(Cipher.ENCRYPT_MODE, key, CHACHA20_PARAM_SPEC_COUNTER_0);
+            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
+            cp.init(Cipher.ENCRYPT_MODE, key, CHACHA20_POLY1305_PARAM_SPEC);
             byte[] cipherText = cp.doFinal(PLAIN_TEXT);
 
-            paramSpec = CHACHA20_PARAM_SPEC_COUNTER_0;
+            paramSpec = cp.getParameters().getParameterSpec(IvParameterSpec.class);
 
             // Verify the text
-            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
+            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
             cp.init(Cipher.DECRYPT_MODE, key, paramSpec);
             byte[] newPlainText = cp.doFinal(cipherText, 0, cipherText.length);
 
@@ -658,20 +748,96 @@ public class BaseTestChaCha20 extends BaseTestCipher implements ChaCha20Constant
     }
 
     //--------------------------------------------------------------------------
+    // Run encrypt/decrypt test using just doFinal calls
+    //
+    @Test
+    public void testChaCha20Poly1305EncryptDecryptDoFinalWithAAD() throws Exception {
+        try {
+            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
+            cp.init(Cipher.ENCRYPT_MODE, key, CHACHA20_POLY1305_PARAM_SPEC);
+            cp.updateAAD(CHACHA20_POLY1305_AAD, 0, CHACHA20_POLY1305_AAD.length);
+            byte[] cipherText = cp.doFinal(PLAIN_TEXT);
+
+            paramSpec = cp.getParameters().getParameterSpec(IvParameterSpec.class);
+
+            // Verify the text
+            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
+            cp.init(Cipher.DECRYPT_MODE, key, paramSpec);
+            cp.updateAAD(CHACHA20_POLY1305_AAD, 0, CHACHA20_POLY1305_AAD.length);
+            byte[] newPlainText = cp.doFinal(cipherText, 0, cipherText.length);
+
+            assertTrue(Arrays.equals(PLAIN_TEXT, newPlainText));
+
+        } catch (Exception e) {
+            fail("Got unexpected exception on encrypt/decrypt...");
+        }
+    }
+
+    //--------------------------------------------------------------------------
+    // Run encrypt/decrypt test using just doFinal calls
+    //
+    @Test
+    public void testChaCha20Poly1305EncryptDecryptDoFinalWithoutAadBadTag() throws Exception {
+        try {
+            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
+            cp.init(Cipher.ENCRYPT_MODE, key, CHACHA20_POLY1305_PARAM_SPEC);
+            byte[] cipherText = cp.doFinal(PLAIN_TEXT);
+            System.arraycopy(BAD_TAG_16, 0, cipherText, cipherText.length - BAD_TAG_16.length,
+                    BAD_TAG_16.length);
+
+            paramSpec = cp.getParameters().getParameterSpec(IvParameterSpec.class);
+
+            // Verify the text
+            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
+            cp.init(Cipher.DECRYPT_MODE, key, paramSpec);
+            cp.doFinal(cipherText, 0, cipherText.length);
+
+            fail("Did not get expected ProviderException: Failure on doFinal(Bad Tag)");
+
+        } catch (Exception e) {
+        }
+    }
+
+    //--------------------------------------------------------------------------
+    // Run encrypt/decrypt test using just doFinal calls
+    //
+    @Test
+    public void testChaCha20Poly1305DecryptDoFinalWithAadNoTag() throws Exception {
+        try {
+
+            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
+            cp.init(Cipher.ENCRYPT_MODE, key, CHACHA20_POLY1305_PARAM_SPEC);
+            cp.updateAAD(CHACHA20_POLY1305_AAD, 0, CHACHA20_POLY1305_AAD.length);
+            cp.doFinal(PLAIN_TEXT);
+
+            paramSpec = cp.getParameters().getParameterSpec(IvParameterSpec.class);
+
+            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
+            cp.init(Cipher.DECRYPT_MODE, key, paramSpec);
+            cp.updateAAD(CHACHA20_POLY1305_AAD, 0, CHACHA20_POLY1305_AAD.length);
+            cp.doFinal();
+
+            fail("Did not get expected IllegalArgumentException(Missing tag on decrypt final)");
+
+        } catch (Exception e) {
+        }
+    }
+
+    //--------------------------------------------------------------------------
     // Run encrypt/decrypt test using just update, empty doFinal calls
     //
     @Test
-    public void testChaCha20EncryptDecryptUpdate() throws Exception {
+    public void testChaCha20Poly1305EncryptDecryptUpdate() throws Exception {
         try {
-            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
-            cp.init(Cipher.ENCRYPT_MODE, key, CHACHA20_PARAM_SPEC_COUNTER_1);
+            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
+            cp.init(Cipher.ENCRYPT_MODE, key, CHACHA20_POLY1305_PARAM_SPEC);
             byte[] cipherText1 = cp.update(PLAIN_TEXT);
             byte[] cipherText2 = cp.doFinal();
 
-            paramSpec = CHACHA20_PARAM_SPEC_COUNTER_1;
+            paramSpec = cp.getParameters().getParameterSpec(IvParameterSpec.class);
 
             // Verify the text
-            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
+            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
             cp.init(Cipher.DECRYPT_MODE, key, paramSpec);
             byte[] newPlainText1 = (cipherText1 == null) ? new byte[0] : cp.update(cipherText1);
             byte[] newPlainText2 = cp.doFinal(cipherText2);
@@ -695,19 +861,19 @@ public class BaseTestChaCha20 extends BaseTestCipher implements ChaCha20Constant
     // Run encrypt/decrypt test with partial update
     //
     @Test
-    public void testChaCha20EncryptDecryptPartialUpdate() throws Exception {
+    public void testChaCha20Poly1305EncryptDecryptPartialUpdate() throws Exception {
         int partialLen = PLAIN_TEXT.length > 10 ? 10 : 1;
 
         try {
-            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
-            cp.init(Cipher.ENCRYPT_MODE, key, CHACHA20_PARAM_SPEC_COUNTER_0);
+            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
+            cp.init(Cipher.ENCRYPT_MODE, key, CHACHA20_POLY1305_PARAM_SPEC);
             byte[] cipherText1 = cp.update(PLAIN_TEXT, 0, partialLen);
             byte[] cipherText2 = cp.doFinal(PLAIN_TEXT, partialLen, PLAIN_TEXT.length - partialLen);
 
-            paramSpec = CHACHA20_PARAM_SPEC_COUNTER_0;
+            paramSpec = cp.getParameters().getParameterSpec(IvParameterSpec.class);
 
             // Verify the text
-            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
+            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
             cp.init(Cipher.DECRYPT_MODE, key, paramSpec);
             byte[] newPlainText1 = (cipherText1 == null) ? new byte[0] : cp.update(cipherText1);
             byte[] newPlainText2 = cp.doFinal(cipherText2);
@@ -731,19 +897,21 @@ public class BaseTestChaCha20 extends BaseTestCipher implements ChaCha20Constant
     // Run encrypt twice using same Cipher instance, same key, but different nonce
     //
     @Test
-    public void testChaCha20ReuseObjectSameKeyDifferentNonce() throws Exception {
+    public void testChaCha20Poly1305ReuseObjectSameKeyDifferentNonce() throws Exception {
         try {
-            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
-            cp.init(Cipher.ENCRYPT_MODE, key, CHACHA20_PARAM_SPEC_COUNTER_0);
+            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
+            cp.init(Cipher.ENCRYPT_MODE, key, CHACHA20_POLY1305_PARAM_SPEC);
             byte[] cipherText1 = cp.doFinal(PLAIN_TEXT);
 
-            paramSpec = CHACHA20_PARAM_SPEC_COUNTER_0;
+            paramSpec = cp.getParameters().getParameterSpec(IvParameterSpec.class);
+
             byte[] newNonce = Arrays.copyOf(NONCE_12_BYTE, NONCE_12_BYTE.length);
             newNonce[0]++;
-            ChaCha20ParameterSpec newNonceSpec = new ChaCha20ParameterSpec(newNonce, COUNTER_0);
+            IvParameterSpec newNonceSpec = new IvParameterSpec(newNonce);
             // Verify that the cipher object can be used to encrypt again with same key different nonce
             cp.init(Cipher.ENCRYPT_MODE, key, newNonceSpec);
             byte[] cipherText2 = cp.doFinal(PLAIN_TEXT);
+
             boolean sameCipher = Arrays.equals(cipherText2, cipherText1);
             assertFalse(sameCipher, "Re-encrypted text with diffent nonce is same");
         } catch (Exception e) {
@@ -755,26 +923,26 @@ public class BaseTestChaCha20 extends BaseTestCipher implements ChaCha20Constant
     // Run encrypt/decrypt test using just doFinal calls (copy-safe)
     //
     @Test
-    public void testChaCha20EncryptDecryptDoFinalCopySafe() throws Exception {
+    public void testChaCha20Poly1305EncryptDecryptDoFinalCopySafe() throws Exception {
         try {
-            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
-            cp.init(Cipher.ENCRYPT_MODE, key, CHACHA20_PARAM_SPEC_COUNTER_0);
+            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
+            cp.init(Cipher.ENCRYPT_MODE, key, CHACHA20_POLY1305_PARAM_SPEC);
 
             byte[] cipherText0 = cp.doFinal(PLAIN_TEXT);
 
             byte[] resultBuffer = Arrays.copyOf(PLAIN_TEXT, cp.getOutputSize(PLAIN_TEXT.length));
-            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
-            cp.init(Cipher.ENCRYPT_MODE, key, CHACHA20_PARAM_SPEC_COUNTER_0);
+            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
+            cp.init(Cipher.ENCRYPT_MODE, key, CHACHA20_POLY1305_PARAM_SPEC);
             int resultLen = cp.doFinal(resultBuffer, 0, PLAIN_TEXT.length, resultBuffer);
             byte[] cipherText = Arrays.copyOf(resultBuffer, resultLen);
 
-            paramSpec = CHACHA20_PARAM_SPEC_COUNTER_0;
+            paramSpec = cp.getParameters().getParameterSpec(IvParameterSpec.class);;
 
             boolean success = Arrays.equals(cipherText, cipherText0);
             assertTrue(success, "Encrypted text does not match expected result");
 
             // Verify the text
-            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
+            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
             cp.init(Cipher.DECRYPT_MODE, key, paramSpec);
             resultBuffer = Arrays.copyOf(cipherText, cipherText.length); //cp.getOutputSize(cipherText.length));
             resultLen = cp.doFinal(resultBuffer, 0, cipherText.length, resultBuffer);
@@ -792,16 +960,16 @@ public class BaseTestChaCha20 extends BaseTestCipher implements ChaCha20Constant
     // Run encrypt/decrypt test using just update, empty doFinal calls (copy-safe)
     //
     @Test
-    public void testChaCha20EncryptDecryptUpdateCopySafe() throws Exception {
+    public void testChaCha20Poly1305EncryptDecryptUpdateCopySafe() throws Exception {
         try {
-            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
-            cp.init(Cipher.ENCRYPT_MODE, key, CHACHA20_PARAM_SPEC_COUNTER_0);
+            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
+            cp.init(Cipher.ENCRYPT_MODE, key, CHACHA20_POLY1305_PARAM_SPEC);
 
             byte[] cipherText0 = cp.doFinal(PLAIN_TEXT);
 
             byte[] cipherText1 = Arrays.copyOf(PLAIN_TEXT, cp.getOutputSize(PLAIN_TEXT.length));
-            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
-            cp.init(Cipher.ENCRYPT_MODE, key, CHACHA20_PARAM_SPEC_COUNTER_0);
+            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
+            cp.init(Cipher.ENCRYPT_MODE, key, CHACHA20_POLY1305_PARAM_SPEC);
             int cipherText1Len = cp.update(cipherText1, 0, PLAIN_TEXT.length, cipherText1);
             byte[] cipherText2 = cp.doFinal();
 
@@ -809,19 +977,18 @@ public class BaseTestChaCha20 extends BaseTestCipher implements ChaCha20Constant
             System.arraycopy(cipherText1, 0, cipherText12, 0, cipherText1Len);
             System.arraycopy(cipherText2, 0, cipherText12, cipherText1Len, cipherText2.length);
 
-            // You can not get the parameters for ChaCha20
-            paramSpec = CHACHA20_PARAM_SPEC_COUNTER_0;
+            paramSpec = cp.getParameters().getParameterSpec(IvParameterSpec.class);
 
             boolean success = Arrays.equals(cipherText12, cipherText0);
             assertTrue(success, "Encrypted text does not match expected result");
 
             // Verify the text
-            cp = Cipher.getInstance(CHACHA20_ALGORITHM, getProviderName());
+            cp = Cipher.getInstance(CHACHA20_POLY1305_ALGORITHM, getProviderName());
             cp.init(Cipher.DECRYPT_MODE, key, paramSpec);
 
             byte[] plainText1 = Arrays.copyOf(cipherText12, cp.getOutputSize(cipherText12.length));
-            int plainText1Len = cp.update(plainText1, 0, cipherText12.length, plainText1);
-            byte[] plainText2 = cp.doFinal();
+            int plainText1Len = cp.update(plainText1, 0, plainText1.length, plainText1);
+            byte[] plainText2 = cp.doFinal(cipherText2);
 
             byte[] plainText12 = new byte[plainText1Len + plainText2.length];
             System.arraycopy(plainText1, 0, plainText12, 0, plainText1Len);
